requirements.txt
astroid==2.3.3
ifaddr==0.1.6
ifcfg==0.20
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
six==1.13.0
tornado==6.0.3
typed-ast==1.4.0
wrapt==1.11.2
zeroconf-py2compat==0.19.4
